TTC now works properly using Minkowski Difference to calculate collisions
