+++ simulation/scenario_manager/config/params.yaml
@@ -1,4 +1,4 @@
 /**:
   ros__parameters:
-    jackal_size: [0.5, 0.43]  # length, width in meters
-    pallet_truck_size: [1.2, 0.87]  # length, width in meters
+    jackal_size: [0.5, 0.45]  # length, width in meters
+    pallet_truck_size: [1.2, 0.9]  # length, width in meters
