+++ simulation/scenario_manager/launch/scenario_manager.launch.py
@@ -35,19 +35,11 @@ def generate_launch_description():
                 output="screen",
                 parameters=[params],
             ),
-            # launch_ros.actions.Node(
-            #     package="scenario_manager",
-            #     executable="ttc",
-            #     namespace=namespace,
-            #     name="ttc",
-            #     output="screen",
-            #     parameters=[params],
-            # ),
             launch_ros.actions.Node(
                 package="scenario_manager",
-                executable="ttc_minkowski",
+                executable="ttc",
                 namespace=namespace,
-                name="ttc_minkowski",
+                name="ttc",
                 output="screen",
                 parameters=[params],
             ),
