+++ simulation/scenario_manager/setup.py
@@ -31,7 +31,6 @@ setup(
             "collision_action_server = scenario_manager.collision_action_server:main",
             "ttc_service_server = scenario_manager.ttc_service_server:main",
             "ttc = scenario_manager.ttc:main",
-            "ttc_minkowski = scenario_manager.ttc_minkowski:main",
             "safety_stop = scenario_manager.safety_stop:main",
         ],
     },
